add custom elm build script
this script is used to produce a `elm-dist/bundle.js` file from an example `src/Main.elm` file.
this is necessary since the default output from `elm-make` will produce a javascript file that expects to execute in browser environment.
this script will wrap the compiled elm code in another code block to defer execution of the code and export a ES module function.